backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
supabase==2.3.0
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1
thefuzz==0.20.0
python-Levenshtein==0.23.0
numpy==1.26.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2